Run workflow on pull requests (#39)
Add Github Action to spin up a test worker when creating a PR to permit interactive testing in a staging environment. Includes a duplicate KV.

At this time, staging workers and KVs must be manually deleted.

Also adds a file upload step to workflow actions for PR tests and merges to master.